+++ src/pages/OnboardingPage/Step/StepService/BoxAllowedService/BoxAllowedService.tsx
