+++ src/pages/OnboardingPage/Step/StepStart/StepStart.tsx
