--- src/pages/OnboardingPage/components/Step/StepService/components/index.ts
@@ -1,3 +1,0 @@
-export { default as BoxAllowedService } from './BoxAllowedService/BoxAllowedService';
-export { default as ButtonService } from './ButtonService/ButtonService';
-export { default as Tabs } from './Tabs/Tabs';
